fix: redid filestructure for pico so that it runs on boot
